+++ Dtos/RiskAssessmentResponseDto.cs
@@ -2,7 +2,7 @@
 {
     public class RiskAssessmentResponseDto
     {
-        public double RiskScore { get; set; }
+        public decimal RiskScore { get; set; }
         public string RiskLevel { get; set; }
         public string Reason { get; set; }
     }
