+++ Services/IRuleEngineService.cs
@@ -4,9 +4,9 @@ namespace VendorRiskScoreAPI.Services
 {
     public interface IRuleEngineService
     {
-        float CalculateFinancialRisk(int financialHealth);
-        float CalculateOperationalRisk(int slaUpTime, int majorIncidents);
-        float CalculateSecurityComplianceRisk(List<VendorSecurityCert> vendorSecurityCerts, Document document);
+        decimal CalculateFinancialRisk(int financialHealth);
+        decimal CalculateOperationalRisk(int slaUpTime, int majorIncidents);
+        decimal CalculateSecurityComplianceRisk(List<VendorSecurityCert> vendorSecurityCerts, Document document);
         Task<string> GetReasons(int slaUpTime, List<VendorSecurityCert> vendorSecurityCerts, Document document);
     }
 }
